+++ libs/base_core/src/data/codec/hash/check/mod.rs
@@ -1,0 +1,14 @@
+// devela_base_core::data::codec::hash::check
+//
+#![doc = crate::_DOC_DATA_CODEC_HASH_CHECK!()]
+//
+
+mod adler; // Adler32
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            adler::*,
+        };
+    }
+}