+++ libs/base_core/src/data/codec/mod.rs
@@ -3,10 +3,14 @@
 #![doc = crate::_DOC_DATA_CODEC!()]
 //
 
+// pub mod frame;
 pub mod hash;
 
 crate::structural_mods! { // _pub_mods
     _pub_mods {
-        pub use super::hash::_all::*;
+        pub use super::{
+            // frame::_all::*,
+            hash::_all::*,
+        };
     }
 }